supertrend/requirements.txt
integrate==1.3.0
pandas==2.0.3
pymongo==4.4.1
pyotp==2.9.0
python_dateutil==2.8.2
Requests==2.31.0
